Store inner and outer bounds respective to agency

+++ osmapp/models.py
@@ -8,8 +8,8 @@ from django.contrib.postgres.fields import ArrayField
 # model for storing bounds specific to the operator
 class FeedBounds(models.Model):
     operator_name = models.CharField(max_length=200)
-    outer_bound = ArrayField(ArrayField(models.DecimalField(max_digits=50, decimal_places=20)), blank=True)
-    inner_bound = ArrayField(ArrayField(models.DecimalField(max_digits=50, decimal_places=20)), blank=True)
+    outer_bound = ArrayField(ArrayField(models.FloatField()), blank=True)
+    inner_bound = ArrayField(ArrayField(models.FloatField()), blank=True)
 
     def __str__(self):  # __unicode__ on Python 2
         return self.operator_name
